fix history

+++ app/src/main/java/com/example/eworkout/training/view/HistoryFragment.kt
@@ -120,7 +120,7 @@ class HistoryFragment : Fragment() {
             "LOADED" -> {
                 binding.num.text = _viewModel.num.toString()
                 binding.min.text = _viewModel.min.toString()
-                binding.numOfExercisesHistory.text = _viewModel.exercises.toString()
+                binding.numOfExercisesHistory.text = _viewModel.exercises.toInt().toString()
                 hideLoading()
             }
             "SET_LOADED" -> {
